refactor(#20): useSearch로 검색로직 파일 분리
