+++ features/search/index.ts
@@ -1,1 +1,2 @@
 export { default as SearchBar } from './ui/SearchBar'
+export { useSearch } from './model/useSearch'
